packaging/karatagor_1.0.0_all/usr/share/karatagor/requirements.txt
PyQt6>=6.9.0
python-vlc>=3.0.21203
mutagen>=1.47.0
pyacoustid>=1.3.0
musicbrainzngs>=0.7.1
pyloudnorm>=0.2.0
requests>=2.32.3
